% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_scoring.R
\name{calcTreeScores}
\alias{calcTreeScores}
\title{Calculate SCHISM fitness scores for trees}
\usage{
calcTreeScores(w_chain, trees, mc.cores = 1)
}
\arguments{
\item{w_chain}{MCMC chain of CCF values, which is the first item in the list returned by \code{clusterSep}}

\item{trees}{list of tibbles, where each tibble contains edges of a tree with columns edge, parent, child}
}
\description{
Calculate SCHISM fitness scores for trees
}
