% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_separately.R
\name{clusterSep}
\alias{clusterSep}
\title{Run MCMC to cluster mutations and estimate CCFs}
\usage{
clusterSep(
  input_data,
  n.iter = 10000,
  n.burn = 1000,
  thin = 10,
  mc.cores = 1,
  inits = list(.RNG.name = "base::Wichmann-Hill", .RNG.seed = 123),
  max_K = 5,
  model_type = "spike_and_slab",
  beta.prior = FALSE,
  drop_zero = FALSE,
  one_box = FALSE
)
}
\arguments{
\item{input_data}{list of input data objects;}

\item{n.iter}{number of iterations to run MCMC}

\item{n.burn}{number of iterations for burn in}

\item{thin}{thinning parameter}

\item{mc.cores}{number of cores for parallelization}

\item{max_K}{maximum number of clusters to assess for each mutation set}

\item{model_type}{hierarchical model type for ("spike_and_slab" or "simple)}

\item{beta.prior}{option to run an initial MCMC chain and use results to specify beta priors for a second MCMC chain}

\item{one_box}{option to run the MCMC chain without using sample presence}
}
\description{
Run MCMC to cluster mutations and estimate CCFs
}
