% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabowmyers.R
\name{enumerateSpanningTrees}
\alias{enumerateSpanningTrees}
\title{Enumerate all spanning trees and filter based on Sum Condition}
\usage{
enumerateSpanningTrees(graph_G, w, sum_filter_thresh = 0.2)
}
\arguments{
\item{graph_G}{tibble of possible edges with columns edge, parent, child}

\item{w}{matrix of CCF values (rows = clusters, columns = samples)}

\item{sum_filter_thresh}{thresh maximum allowed violation of Sum Condition (default = 0.2)}
}
\description{
Enumerate all spanning trees and filter based on Sum Condition
}
