% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-gabowmyers.R
\name{enumerateSpanningTreesModified}
\alias{enumerateSpanningTreesModified}
\title{Enumerate all spanning trees using modified Gabow-Myers}
\usage{
enumerateSpanningTreesModified(graph_G, mcf, purity, sum_filter_thresh = 0.2)
}
\arguments{
\item{graph_G}{tibble of possible edges with columns edge, parent, child}

\item{sum_filter_thresh}{thresh maximum allowed violation of Sum Condition (default = 0.2)}

\item{w}{matrix of CCF values (rows = clusters, columns = samples)}
}
\description{
Enumerate all spanning trees using modified Gabow-Myers
}
