% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{estimateCCFs}
\alias{estimateCCFs}
\title{Determine the most probable cluster CCF values by taking the mode of the posterior distributions}
\usage{
estimateCCFs(w_chain)
}
\arguments{
\item{w_chain}{MCMC chain of CCF values, which is the first item in the list returned by \code{clusterSep}}
}
\value{
matrix of estimated cluster CCFs
}
\description{
Determine the most probable cluster CCF values by taking the mode of the posterior distributions
}
