% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{estimateClusterAssignments}
\alias{estimateClusterAssignments}
\title{Determine most probable mutation cluster assignments by taking those with highest posterior probability}
\usage{
estimateClusterAssignments(z_chain)
}
\arguments{
\item{z_chain}{MCMC chain of mutation cluster assignment values, which is the second item in the list returned by \code{clusterSep}}
}
\description{
Determine most probable mutation cluster assignments by taking those with highest posterior probability
}
