% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{estimateICN}
\alias{estimateICN}
\title{Determine most probable integer assignments by taking those with highest posterior probability}
\usage{
estimateICN(icn_chain)
}
\arguments{
\item{icn_chain}{MCMC chain of integer copy number assignment values}
}
\description{
Determine most probable integer assignments by taking those with highest posterior probability
}
