% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{estimateMCFs}
\alias{estimateMCFs}
\title{Determine the most probable cluster MCF values by taking the mean of the posterior distributions}
\usage{
estimateMCFs(mcf_chain)
}
\arguments{
\item{mcf_chain}{MCMC chain of mCF values}
}
\value{
matrix of estimated cluster MCFs
}
\description{
Determine the most probable cluster MCF values by taking the mean of the posterior distributions
}
