% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{estimateMultiplicity}
\alias{estimateMultiplicity}
\title{Determine most probable mutation cluster assignments by taking those with highest posterior probability}
\usage{
estimateMultiplicity(m_chain)
}
\arguments{
\item{m_chain}{MCMC chain of multiplicity assignment values}
}
\description{
Determine most probable mutation cluster assignments by taking those with highest posterior probability
}
