% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importFiles}
\alias{importFiles}
\title{read input data file and store in required format}
\usage{
importFiles(
  mutation_file,
  outputDir = NULL,
  alt_reads_thresh = 0,
  vaf_thresh = 0
)
}
\arguments{
\item{mutation_file}{mutation data file; see inst/extdata/examples/*_snv.csv for examples}

\item{outputDir}{output directory for saving output data}
}
\description{
read input data file and store in required format
}
