% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importMutationFileOnly}
\alias{importMutationFileOnly}
\title{import mutation file}
\usage{
importMutationFileOnly(mutation_file, alt_reads_thresh = 0, vaf_thresh = 0)
}
\description{
import mutation file
}
