% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-main.R
\name{mcmcMain}
\alias{mcmcMain}
\title{run PICTograph in an automated pipeline}
\usage{
mcmcMain(
  mutation_file,
  outputDir = NULL,
  sample_presence = TRUE,
  score = "silhouette",
  max_K = 10,
  min_mutation_per_cluster = 5,
  cluster_diff_thresh = 0.05,
  n.iter = 5000,
  n.burn = 1000,
  thin = 10,
  mc.cores = 8,
  inits = list(.RNG.name = "base::Wichmann-Hill", .RNG.seed = 123),
  alt_reads_thresh = 0,
  vaf_thresh = 0
)
}
\arguments{
\item{mutation_file}{a csv file that include information for SSMs.}

\item{outputDir}{output directory for saving all files.}

\item{sample_presence}{whether to use sample presence to separate the mutations. Not applicable if dual_model is set to FALSE and a copy number file is provided.}

\item{score}{scoring function to estimate the number of clusters. silhouette or BIC.}

\item{max_K}{user defined maximum number of clusters.}

\item{min_mutation_per_cluster}{minumum number of mutations in each cluster.}

\item{cluster_diff_thresh}{threshold to merge two clusters.}

\item{n.iter}{number of iterations by JAGS.}

\item{n.burn}{number of burns by JAGS.}

\item{thin}{number of thin by JAGS.}

\item{mc.cores}{number of cores to use for parallel computing; not applicable to windows.}

\item{inits}{additional parameters by JAGS.}
}
\description{
run MCMC chains to  infer the clonal evolution of tumors from single or multi-region sequencing data. 
This function automatically runs a pipeline of the tool. It models uncertainty of mutation cellular 
fraction (MCF) in small somatic mutations (SSMs) and copy number alterations (CNAs), assigning SSMs 
and CNAs to subclones using a Bayesian hierarchical model, and reconstruct tumor evolutionary trees 
that are constrained based on principles of lineage precedence, sum condition, and optionally by 
sample-presence.
}
