% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{mergeSetChains}
\alias{mergeSetChains}
\title{Relabel chains for all sets and merge}
\usage{
mergeSetChains(best_set_chains, indata)
}
\arguments{
\item{best_set_chains}{List of lists of MCMC chains (mcf_chain, z_chain, ystar_chain) for each mutation set}

\item{indata}{List of input data objects (same as provided to clusterSep)}
}
\description{
Relabel chains for all sets and merge
}
