% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-main.R
\name{plotAllTrees}
\alias{plotAllTrees}
\title{Plot all trees with the highest scores}
\usage{
plotAllTrees(outputDir, scores, all_spanning_trees, mcfTable, data)
}
\description{
Plot all trees with the highest scores
}
