% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBIC}
\alias{plotBIC}
\title{Plot BIC for all mutation sets}
\usage{
plotBIC(all_set_results, Sample_ID = NULL)
}
\arguments{
\item{all_set_results}{List of MCMC results for each mutation set; returned by \code{clusterSep}}

\item{SampleID}{(Optional) Vector of sample IDs for labeling purposes. Same order as supplied as input data (e.g. indata$Sample_ID)}
}
\description{
Plot BIC for all mutation sets
}
