% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotChainsCCF}
\alias{plotChainsCCF}
\title{Plot CCF chain trace}
\usage{
plotChainsCCF(w_chain)
}
\arguments{
\item{w_chain}{MCMC chain of CCF values, which is the first item in the list returned by \code{clusterSep}}
}
\description{
Plot CCF chain trace
}
