% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_plot_clusters.R
\name{plotChainsMCF}
\alias{plotChainsMCF}
\title{Plot CCF chain trace}
\usage{
plotChainsMCF(mcf_chain)
}
\arguments{
\item{mcf_chain}{MCMC chain of CCF values, which is the first item in the list returned by \code{mergeSetChains}}
}
\description{
Plot CCF chain trace
}
