% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterAssignmentProb}
\alias{plotClusterAssignmentProb}
\title{Plot probabilities of mutation cluster assignments}
\usage{
plotClusterAssignmentProb(z_chain)
}
\arguments{
\item{z_chain}{MCMC chain of mutation cluster assignment values, which is the second item in the list returned by \code{clusterSep}}
}
\description{
Plot probabilities of mutation cluster assignments
}
