% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterAssignmentProbVertical}
\alias{plotClusterAssignmentProbVertical}
\title{Plot probabilities of mutation cluster assignments - vertical}
\usage{
plotClusterAssignmentProbVertical(
  z_chain,
  w_chain,
  filter_thresh = 0.05,
  MutID = NULL,
  SampleID = NULL
)
}
\arguments{
\item{z_chain}{MCMC chain of mutation cluster assignment values, which is the second item in the list returned by \code{clusterSep}}

\item{filter_thresh}{Lowest posterior probability to include cluster assignment. Default value is 0.05 (inclusive)}

\item{MutID}{(Optional) Vector of mutation IDs for labeling purposes. Same order as supplied as input data (e.g. indata$Mut_ID)}
}
\description{
Plot probabilities of mutation cluster assignments - vertical
}
