% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDensityCCF}
\alias{plotDensityCCF}
\title{Plot cluster CCF posterior distributions}
\usage{
plotDensityCCF(w_chain)
}
\arguments{
\item{w_chain}{MCMC chain of CCF values, which is the first item in the list returned by \code{clusterSep}}
}
\description{
Plot cluster CCF posterior distributions
}
