% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-plot.R
\name{plotEnsembleTree}
\alias{plotEnsembleTree}
\title{Plot ensemble tree}
\usage{
plotEnsembleTree(trees, palette = viridis::viridis)
}
\arguments{
\item{trees}{list of tibbles of edges, each with columns edge, parent, child}
}
\description{
Plot ensemble tree
}
