% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_plot_clusters.R
\name{plotMCFViolin}
\alias{plotMCFViolin}
\title{Plot cluster CCF posterior distributions as violin plots}
\usage{
plotMCFViolin(mcf_chain, z_chain = NULL, indata = NULL)
}
\arguments{
\item{mcf_chain}{MCMC chain of CCF values}

\item{z_chain}{(Optional) MCMC chain of mutation cluster assignment values. If provided, cluster names will show the number of mutations assigned in brackets}

\item{indata}{(Optional) List of input data items}
}
\description{
Plot cluster CCF posterior distributions as violin plots
}
