% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclone-proportion.R
\name{plotSubcloneBar}
\alias{plotSubcloneBar}
\title{Plot subclone proportions in each sample as stacked bar chart}
\usage{
plotSubcloneBar(
  subclone_props,
  palette = viridis::viridis,
  sample_names = NULL,
  label_cluster = FALSE
)
}
\arguments{
\item{subclone_props}{matrix of subclone proportions (returned from \code{calcSubcloneProportions})}

\item{sample_names}{(Optional) Vector of sample names. Should be in the order of columns of subclone_props}

\item{label_cluster}{(Default FALSE) Whether to add cluster label to text on stacked bar}
}
\description{
Plot subclone proportions in each sample as stacked bar chart
}
