% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-plot.R
\name{plotTree}
\alias{plotTree}
\title{Plot single tree}
\usage{
plotTree(edges, palette = viridis::viridis)
}
\arguments{
\item{edges}{tibble of edges with columns edge, parent, child}
}
\description{
Plot single tree
}
