% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-gabowmyers.R
\name{prepareGraph}
\alias{prepareGraph}
\title{Create tibble of possible edges from CCF values based on w_mat only}
\usage{
prepareGraph(mcf_mat, thresh)
}
\arguments{
\item{w}{matrix of CCF values (rows = clusters, columns = samples)}
}
\value{
graph_G tibble of possible edges with columns edge, parent, child
}
\description{
Create tibble of possible edges from CCF values based on w_mat only
}
