% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabowmyers.R
\name{prepareGraphForGabowMyers}
\alias{prepareGraphForGabowMyers}
\title{Create tibble of possible edges from CCF values based on sample-presence}
\usage{
prepareGraphForGabowMyers(w, zero.thresh = 0.01)
}
\arguments{
\item{w}{matrix of CCF values (rows = clusters, columns = samples)}

\item{zero.thresh}{lower CCF threshold for consider a mutation to be present}
}
\value{
graph_G tibble of possible edges with columns edge, parent, child
}
\description{
Create tibble of possible edges from CCF values based on sample-presence
}
