% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-presence.R
\name{separateMutationsBySamplePresence}
\alias{separateMutationsBySamplePresence}
\title{sample presence for MCMC}
\usage{
separateMutationsBySamplePresence(input_data)
}
\description{
sample presence for MCMC
}
