% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{writeClusterCCFsTable}
\alias{writeClusterCCFsTable}
\title{Determine the most probable cluster CCF values by taking the mode of the posterior distributions}
\usage{
writeClusterCCFsTable(w_chain, Sample_ID)
}
\arguments{
\item{w_chain}{MCMC chain of CCF values, which is the first item in the list returned by \code{clusterSep}}

\item{Sample_ID}{Vector of sample IDs, same order as provided as input data (e.g. indata$Sample_ID)}
}
\value{
A tibble of estimated cluster CCFs in each sample
}
\description{
Determine the most probable cluster CCF values by taking the mode of the posterior distributions
}
