% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{writeIcnTable}
\alias{writeIcnTable}
\title{Determine most probable integer copy number by taking those with highest posterior probability.}
\usage{
writeIcnTable(icn_chain, Mut_ID = NULL)
}
\arguments{
\item{icn_chain}{MCMC chain of integer copy number}

\item{Mut_ID}{Vector of mutation IDs, same order as provided as input data (e.g. indata$Mut_ID)}
}
\value{
A tibble listing mutation IDs and their cluster assignments
}
\description{
Determine most probable integer copy number by taking those with highest posterior probability.
}
