% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{writeMultiplicityTable}
\alias{writeMultiplicityTable}
\title{Determine most probable multiplicity assignments by taking those with highest posterior probability.}
\usage{
writeMultiplicityTable(m_chain, Mut_ID = NULL)
}
\arguments{
\item{m_chain}{MCMC chain of mutation cluster assignment values}

\item{Mut_ID}{Vector of mutation IDs, same order as provided as input data (e.g. indata$Mut_ID)}
}
\value{
A tibble listing mutation IDs and their cluster assignments
}
\description{
Determine most probable multiplicity assignments by taking those with highest posterior probability.
}
