% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-BIC.R
\name{writeSetKTable}
\alias{writeSetKTable}
\title{Make table listing possible choices of K (minimum BIC and elbow of BIC plot) for each mutation set}
\usage{
writeSetKTable(all_set_results, sample_names = NULL)
}
\arguments{
\item{all_set_results}{List of MCMC results for each mutation set; returned by \code{clusterSep}}

\item{sample_names}{(Optional) Vector of sample IDs, same order as provided as input data (e.g. indata$Sample_ID)}
}
\description{
Make table listing possible choices of K (minimum BIC and elbow of BIC plot) for each mutation set
}
