% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclone-proportion.R
\name{calcSubcloneProportions}
\alias{calcSubcloneProportions}
\title{Calculate proportions of subclones in each sample (assumes CCFs comply with lineage precedence and sum condition)}
\usage{
calcSubcloneProportions(w_mat, tree_edges)
}
\arguments{
\item{w_mat}{Matrix of CCF estimates (from \code{estimateCCFs})}

\item{tree_edges}{Tibble of tree edges with columns edge, parent, and child}
}
\description{
Calculate proportions of subclones in each sample (assumes CCFs comply with lineage precedence and sum condition)
}
