% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_separately.R
\name{clusterSep}
\alias{clusterSep}
\title{Run MCMC to cluster mutations and estimate CCFs}
\usage{
clusterSep(
  input_data,
  n.iter = 10000,
  n.burn = 1000,
  thin = 10,
  mc.cores = 1,
  inits = list(.RNG.name = "base::Wichmann-Hill", .RNG.seed = 123),
  params = c("z", "w"),
  max_K = 10,
  opt_K = "min_BIC"
)
}
\arguments{
\item{input_data}{list of input data objects;}

\item{opt_K}{strategy for choosing optimum K ("min_BIC" or "elbow1")}
}
\description{
Run MCMC to cluster mutations and estimate CCFs
}
