% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{collectBestKChains}
\alias{collectBestKChains}
\title{Collect chains for best K of each mutation set}
\usage{
collectBestKChains(all_set_results, chosen_K = NULL)
}
\arguments{
\item{all_set_results}{List of MCMC results for each mutation set}

\item{chosen_K}{(Optional) Vector of K to choose for each mutation set, in the same order as all_set_results. If left blank, function will select best K automatically selected by \code{clusterSep}}
}
\description{
Collect chains for best K of each mutation set
}
