% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{estimateClusterAssignments}
\alias{estimateClusterAssignments}
\title{Determine most probable mutation cluster assignments by taking those with highest posterior probability}
\usage{
estimateClusterAssignments(z_chain)
}
\arguments{
\item{z_chain}{MCMC chain of mutation cluster assignment values}
}
\description{
Determine most probable mutation cluster assignments by taking those with highest posterior probability
}
