% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-gabowmyers.R
\name{filterEdgesBasedOnCCFs}
\alias{filterEdgesBasedOnCCFs}
\title{Filter possible edges based on lineage precedence}
\usage{
filterEdgesBasedOnCCFs(graph_G, mcf, thresh = 0.1)
}
\arguments{
\item{graph_G}{tibble of possible edges with columns edge, parent, child}

\item{thresh}{maximum allowed violation of lineage precedence (default = 0.1)}

\item{w}{matrix of CCF values (rows = clusters, columns = samples)}
}
\description{
Filter possible edges based on lineage precedence
}
