% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-gabowmyers.R
\name{generateAllTrees}
\alias{generateAllTrees}
\title{Enumerate all spanning trees using modified Gabow-Myers wrapper}
\usage{
generateAllTrees(
  mcf,
  purity,
  lineage_precedence_thresh = 0.1,
  sum_filter_thresh = 0.2
)
}
\arguments{
\item{lineage_precedence_thresh}{maximum allowed violation of lineage precedence (default = 0.1)}

\item{sum_filter_thresh}{thresh maximum allowed violation of Sum Condition (default = 0.2)}

\item{w}{matrix of CCF values (rows = clusters, columns = samples)}
}
\description{
Enumerate all spanning trees using modified Gabow-Myers wrapper
}
