% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importCSV}
\alias{importCSV}
\title{read input data file and store in PICTOGRAPH input format}
\usage{
importCSV(inputFile, alt_reads_thresh = 6, vaf_thresh = 0.02)
}
\arguments{
\item{input_file}{input data file;}
}
\description{
read input data file and store in PICTOGRAPH input format
}
