% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCCFViolin}
\alias{plotCCFViolin}
\title{Plot cluster CCF posterior distributions as violin plots}
\usage{
plotCCFViolin(w_chain, indata)
}
\arguments{
\item{w_chain}{MCMC chain of CCF values, which is the first item in the list returned by \code{clusterSep}}

\item{indata}{List of input data items}
}
\description{
Plot cluster CCF posterior distributions as violin plots
}
