% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclone-proportion.R
\name{plotSubclonePie}
\alias{plotSubclonePie}
\title{Plot pie charts for subclone proportions in each sample}
\usage{
plotSubclonePie(
  subclone_props,
  palette = viridis::viridis,
  sample_names = NULL,
  title_size = 16,
  legend_size = 10
)
}
\arguments{
\item{subclone_props}{matrix of subclone proportions (returned from \code{calcSubcloneProportions})}

\item{sample_names}{(Optional) Vector of sample names. Should be in the order of columns of subclone_props}
}
\description{
Plot pie charts for subclone proportions in each sample
}
