% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-process-clusters.R
\name{writeClusterAssignmentsTable}
\alias{writeClusterAssignmentsTable}
\title{Determine most probable mutation cluster assignments by taking those with highest posterior probability.}
\usage{
writeClusterAssignmentsTable(
  z_chain,
  mcf_chain = NULL,
  cncf = NULL,
  Mut_ID = NULL
)
}
\arguments{
\item{z_chain}{MCMC chain of mutation cluster assignment values}

\item{Mut_ID}{Vector of mutation IDs, same order as provided as input data (e.g. indata$Mut_ID)}
}
\value{
A tibble listing mutation IDs and their cluster assignments
}
\description{
Determine most probable mutation cluster assignments by taking those with highest posterior probability.
}
